app/requirements.txt
Flask-Cors 
requests 
cbor2 
pytz 
flask 
pyignite 
tinyec 
pycryptodome  
config 
flask_api 
cbor_diag 
validators 
Flask-Session 
jsonschema 
flask==2.3.3 
werkzeug==2.3.7 
pycose 
cwt
pyjwt
schedule
pillow
gunicorn
git+https://github.com/devisefutures/pyMDOC-CBOR.git@cert_arg
git+https://github.com/openwallet-foundation-labs/sd-jwt-python.git
git+https://github.com/Scytales-DeviseFutures/idpy-oidc.git@Scy
git+https://github.com/Scytales-DeviseFutures/openid4v.git@Scy
git+https://github.com/rohe/fedservice.git@107799de02ce282f8c9feaf85e4ba6c88a1d76ab
git+https://github.com/rohe/idpy-sdjwt.git
